requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-redis==5.4.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.7
python-decouple==3.8
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0